Take care of more cases of HDL cards, devices and containers. Add comprehensive tristate support for device workers, and tieoffs. Deal with platforms that don't use all signals in a slot. Deal with platforms that don't support all of a device's signals. Eliminate the redundant names and xml files for "base" platform configurations. Tieoff signals for devices not being used, to preserve a singls constraints file. Add si5351 clock gen support with proper tristate support and using opencores. Update ocpigen ethers command to include IP addresses and systemID
hdl/assemblies/empty/Makefile: include default containers
hdl/cards/specs/fmc_lpc.xml: remove carrier-driven signals from the slot definition
-- This signals will not be driven by FPGAs on the carrier
hdl/devices/Makefile: add si5351 device that just uses opencores
hdl/platforms/alst4/Makefile: remote default base configuration - they are automatic now
hdl/platforms/alst4x/Makefile: ditto
hdl/platforms/alst4x/alst4x_base.xml: removed
hdl/platforms/isim_pf/Makefile: ditto
hdl/platforms/zed/Makefile: ditto, but keep in the list
hdl/platforms/zed/zed.xml: indicate slot signals that this platform doesn't connect to
hdl/platforms/zed_zipper/Makefile: add configurations (base and zipper)
hdl/platforms/zed_zipper/base.xml: removed
hdl/platforms/zed_zipper/zed_zipper.ucf: add full spec for tristate IOBUFs
hdl/platforms/zed_zipper/zed_zipper.vhd: enable it to build now
hdl/platforms/zed_zipper/zed_zipper.xml: indicate device signals that are not pinned-out
hdl/primitives/bsv/bsv_pkg.vhd: add TriState component decl
hdl/primitives/util/util_pkg.vhd: add TSINOUT_1 and TSINOUT_N component declarations
runtime/hdl/src/ocpihdl_main.cxx: print all interfaces and their IP addresses for "ethers"
runtime/util/misc/include/OcpiUtilMisc.h: add getSystemId declaration
hdl/devices/Makefile: remove dram device workers until inouts with buried IOBUFs are working
hdl/devices/flash.hdl/flash.v: update old verilog for new inout scheme
os/linux/src/OcpiOsEther.cxx: don't consider the loopback as an ethernet interface
runtime/hdl/src/ocpihdl_main.cxx: print IP address for each ethernet if available
runtime/util/misc/src/OcpiUtilMisc.cxx: new getSystemId function: first ether mac address
tools/cdk/include/hdl/hdl-assembly.mk: remove redundant platform prefix from configs
tools/cdk/include/hdl/hdl-platform.mk: auto-generate the base configuration xml
-- also remove the redundant platform prefix from these base configurations
tools/cdk/include/hdl/modelsim.mk: ensure that modelsim finds verilog includes in the CWD
tools/ocpigen/include/assembly.h: change pointer to reference
tools/ocpigen/include/hdl-container.h: record which device signals are connected in container
-- enable container-level tieoffs for slot and platform signals that are not used.
tools/ocpigen/include/hdl-device.h: remove unused members
tools/ocpigen/include/ocpigen.h: add support for tristate signals in device workers
-- device workers do not use inout signals, but container code gen creates buffers
tools/ocpigen/include/wip.h: add container tristate tieoff support, and better "worker parent"
tools/ocpigen/src/assembly.cxx: supply true assembly parent to instances
tools/ocpigen/src/hdl-assembly.cxx: tristate support
tools/ocpigen/src/hdl-config.cxx: typo
tools/ocpigen/src/hdl-container.cxx: default platform config prefix removed
-- add more top level tieoff support
-- respect platforms that don't use some slot signals
-- respect device signals that are not in the platform or card
tools/ocpigen/src/hdl-device.cxx: fix bug with wrong parent assembly worker..
tools/ocpigen/src/hdl-worker.cxx: tristate support
tools/ocpigen/src/hdl.cxx: tristate support
tools/ocpigen/src/parse.cxx: default config language to vhdl
tools/ocpigen/src/wci.cxx: remove unused signals that are now in the "raw" record
hdl/platforms/ml605/ml605_base.xml: remove unnecessary base config
runtime/remote/src/RemoteLauncher.cxx: warning
hdl/assemblies/empty/dc.xml: use isim, and don't ask for subdevices since they are automatic
hdl/platforms/ml605/Makefile: use proper name for automatic base config
hdl/devices/si5351.hdl: add this worker now that it uses opencores
hdl/platforms/specs/base.xml: the default base configuration <HdlConfig/>
hdl/platforms/zed_zipper/zipper.xml: the config with these devices in this platform
hdl/primitives/util/TSINOUT_1.vhd: new tristate single output
hdl/primitives/util/TSINOUT_N.vhd: new tristate vector output

+++ runtime/remote/src/RemoteLauncher.cxx
@@ -371,7 +371,7 @@ launch(Launcher::Instances &instances, Launcher::Connections &connections) {
 	artN = arti->second;
       emitInstance(i->m_name.c_str(), contN, artN, *i,
 		   i->m_slave && &i->m_slave->m_container->launcher() == this ?
-		   m_instanceMap[i->m_slave - &instances[0]] : -1);
+		   (int)(m_instanceMap[i->m_slave - &instances[0]]) : -1);
     }
   m_artifacts.resize(nArtifacts);
   unsigned nConnections = 0;
